Remove GEE dependency and replace with STAC module
Now that APR-179 confirmed the STAC path is live in the app:
- Delete src/onkia/satellite_lst.py (GEE-based satellite module)
- Delete tests/test_satellite_lst.py (tests for the old GEE module)
- Move shared dataclasses, constants, and utility functions (LSTObservation,
  LSTHistoryPoint, NDCIObservation, LAKE_ACRES, celsius_to_fahrenheit, etc.)
  directly into satellite_lst_stac.py, removing the cross-module import
- Update test_satellite_lst_stac.py to import those types from satellite_lst_stac
- Remove earthengine-api>=0.1.370 from pyproject.toml and requirements.txt

All 218 relevant tests pass. Fixes APR-180.
