+++ pyproject.toml
@@ -24,7 +24,6 @@ folium = ">=0.14.0"
 xarray = ">=2023.0"
 pyarrow = ">=12.0"
 scipy = ">=1.10"
-earthengine-api = ">=0.1.370"
 pystac-client = ">=0.8"
 planetary-computer = ">=1.0"
 rioxarray = ">=0.18"
