+++ requirements.txt
@@ -11,7 +11,6 @@ great-expectations>=0.18.0
 xarray>=2023.0
 pyarrow>=12.0
 scipy>=1.10
-earthengine-api>=0.1.370
 pystac-client>=0.8
 planetary-computer>=1.0
 rioxarray>=0.18
